src/requirements.txt
streamlit
gradio
# google SDK
google-generativeai # old version Google gnerative AI SDK
google-genai # new version Google gnerative AI SDK
python-dotenv
pandas
numpy
scipy
scikit-learn
matplotlib
seaborn
plotly
plotly-express
huggingface_hub

pytubefix
opencv-python
pillow

# FAST API 
fastapi
uvicorn[standard]
nest-asyncio
aiohttp
pydantic

# Message Queue
rq
redis